pyproject.toml
[tool.poetry]
name = "desafio-vaga-meta"
version = "0.1.0"
description = ""
authors = ["DenysMenfredy <denys.menfredy7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
PyMySQL = "^1.0.2"
pytest = "^7.1.2"
alembic = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-xdist = "^2.5.0"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
